Tidied Up CellBuilder, and Map.
Added board, length, and width to Map.
Added Iterator to the Map to cycle through the Cells, width first, then row.

Removed Error checking from CellBuilder::build_string, as it was causing a partially empty Vec to be returned. The Error will be in the Map later on anyway...

More testing and 100% less memes